ndk: Add `_ = !0` catch-all constant to all `bitflags`, allowing them to be extended
Android may return values in `bitflags` that are not yet mapped or
recognized in the `ndk` crate.  This makes functions like `all()` and
`!` behave unexpectedly when they truncate or limit their operation to
known bits exclusively.

A more likable solution is disabling functions that are susceptible to
this, as was done in the `ash` crate.