ndk/event: Implement `SourceClass` `bitflag` and provide `Source::class()` getter (#458)
* ndk/event: Implement `SourceClass` `bitflag` and provide `Source::class()` getter

This `enum` was never `pub` even though it serves a slight purpose to
more easily classify various event sources, which is accounted for in
the bitmask-like value of `AINPUT_SOURCE_XXX`.  Even though every source
currently only appears to use a single class value, the class variants
all describe an individual bit and should be treated as a bitflags type.

* ndk: Add `_ = !0` catch-all constant to all `bitflags`, allowing them to be extended

Android may return values in `bitflags` that are not yet mapped or
recognized in the `ndk` crate.  This makes functions like `all()` and
`!` behave unexpectedly when they truncate or limit their operation to
known bits exclusively.

A more likable solution is disabling functions that are susceptible to
this, as was done in the `ash` crate.